// src/app/interfaces/interfaces.ts
export interface User {

  id: number,
  name: string,
  password: string,
  email:  string,
  check?: boolean,
  country: Countries,
  city: string,
}